# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SkyGuard OS drone-detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The protected property location and geofence radius
 */
export interface HomeConfig {
  propertyName: string;
  lat: number;
  lng: number;
  geofenceRadiusMeters: number;
  updatedAt: string;
}

export interface HomeConfigInput {
  /** @minLength 1 */
  propertyName: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  lat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  lng: number;
  /** @minimum 1 */
  geofenceRadiusMeters: number;
}

/**
 * A single telemetry sample reported by the roof bridge (HackRF/Raspberry Pi)
 */
export interface DetectionInput {
  /** Stable identifier for the drone session (e.g. DroneID serial or MAC-derived id) */
  droneId: string;
  /** @nullable */
  model?: string | null;
  /**
     * e.g. OcuSync, DroneID, FPV
     * @nullable
     */
  signalType?: string | null;
  /**
     * @minimum -90
     * @maximum 90
     */
  lat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  lng: number;
  /** @nullable */
  altitudeM?: number | null;
  /** @nullable */
  speedKmh?: number | null;
  /** @nullable */
  headingDeg?: number | null;
  /** @nullable */
  rssiDbm?: number | null;
  /** @nullable */
  pilotLat?: number | null;
  /** @nullable */
  pilotLng?: number | null;
  /**
     * Defaults to server receipt time if omitted
     * @nullable
     */
  timestamp?: string | null;
}

export interface Detection {
  id: number;
  droneId: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  signalType?: string | null;
  lat: number;
  lng: number;
  /** @nullable */
  altitudeM?: number | null;
  /** @nullable */
  speedKmh?: number | null;
  /** @nullable */
  headingDeg?: number | null;
  /** @nullable */
  rssiDbm?: number | null;
  /** @nullable */
  pilotLat?: number | null;
  /** @nullable */
  pilotLng?: number | null;
  distanceFromHomeM: number;
  timestamp: string;
  createdAt: string;
}

export interface DroneTrackPoint {
  lat: number;
  lng: number;
  /** @nullable */
  altitudeM?: number | null;
  timestamp: string;
}

/**
 * An actively tracked drone with its recent flight path
 */
export interface DroneTrack {
  droneId: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  signalType?: string | null;
  firstSeenAt: string;
  lastSeenAt: string;
  lat: number;
  lng: number;
  /** @nullable */
  altitudeM?: number | null;
  /** @nullable */
  speedKmh?: number | null;
  /** @nullable */
  headingDeg?: number | null;
  /** @nullable */
  rssiDbm?: number | null;
  /** @nullable */
  pilotLat?: number | null;
  /** @nullable */
  pilotLng?: number | null;
  distanceFromHomeM: number;
  alarmActive: boolean;
  path: DroneTrackPoint[];
}

export interface FlightSummary {
  droneId: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  signalType?: string | null;
  firstSeenAt: string;
  lastSeenAt: string;
  minDistanceM: number;
  /** @nullable */
  maxAltitudeM?: number | null;
  /** @nullable */
  maxSpeedKmh?: number | null;
  pointCount: number;
  stillActive: boolean;
}

export interface IngestStatus {
  /** True if a detection was received within the last 15 seconds */
  connected: boolean;
  /** @nullable */
  lastIngestAt: string | null;
  activeDroneCount: number;
  detectionsToday: number;
}

/**
 * A registered field hardware device (e.g. Raspberry Pi + HackRF bridge)
 */
export interface Device {
  id: number;
  name: string;
  /** Non-secret prefix of the API key, for identification only */
  apiKeyPrefix: string;
  revoked: boolean;
  createdAt: string;
  /** @nullable */
  lastSeenAt: string | null;
}

export interface CreateDeviceInput {
  /** @minLength 1 */
  name: string;
}

export type DeviceWithKey = Device & {
  /** Plaintext API key — shown only once, at creation time */
  apiKey: string;
};

export type ListFlightHistoryParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};