# lib/api-zod/src/generated/types/createDeviceInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SkyGuard OS drone-detection API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDeviceInput {
  /** @minLength 1 */
  name: string;
}