# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SkyGuard OS drone-detection API
 * OpenAPI spec version: 0.1.0
 */

/**
 * A registered field hardware device (e.g. Raspberry Pi + HackRF bridge)
 */
export interface Device {
  id: number;
  name: string;
  /** Non-secret prefix of the API key, for identification only */
  apiKeyPrefix: string;
  revoked: boolean;
  createdAt: Date;
  /** @nullable */
  lastSeenAt: Date | null;
}