# lib/api-zod/src/generated/types/deviceWithKey.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SkyGuard OS drone-detection API
 * OpenAPI spec version: 0.1.0
 */
import type { Device } from './device';

export type DeviceWithKey = Device & {
  /** Plaintext API key — shown only once, at creation time */
  apiKey: string;
};